+++ bmc-integration/src/main/java/com/oracle/bmc/integration/model/UpdateIntegrationInstanceDetails.java
@@ -318,6 +318,7 @@ public final class UpdateIntegrationInstanceDetails
         Enterprise("ENTERPRISE"),
         Standardx("STANDARDX"),
         Enterprisex("ENTERPRISEX"),
+        Healthcare("HEALTHCARE"),
         ;
 
         private final String value;