+++ bmc-networkloadbalancer/src/main/java/com/oracle/bmc/networkloadbalancer/model/CreateNetworkLoadBalancerDetails.java
@@ -6,8 +6,7 @@ package com.oracle.bmc.networkloadbalancer.model;
 
 /**
  * The properties that define a network load balancer. For more information, see [Managing a network
- * load
- * balancer](https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/NetworkLoadBalancers/network-load-balancer-management.htm).
+ * load balancer](https://docs.cloud.oracle.com/Content/Balance/Tasks/managingloadbalancer.htm).
  *
  * <p>To use any of the API operations, you must be authorized in an IAM policy. If you are not
  * authorized, then contact an administrator. If you are an administrator who writes policies to
@@ -194,8 +193,8 @@ public final class CreateNetworkLoadBalancerDetails
          * rules](https://docs.cloud.oracle.com/Content/network/Concepts/securitylists.htm) for your
          * virtual cloud network. For more information about public and private network load
          * balancers, see [How Network Load Balancing
-         * Works](https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/overview.htm). This
-         * value is true by default.
+         * Works](https://docs.cloud.oracle.com/Content/Balance/Concepts/balanceoverview.htm#how-network-load-balancing-works).
+         * This value is true by default.
          *
          * <p>Example: {@code true}
          */
@@ -214,8 +213,8 @@ public final class CreateNetworkLoadBalancerDetails
          * rules](https://docs.cloud.oracle.com/Content/network/Concepts/securitylists.htm) for your
          * virtual cloud network. For more information about public and private network load
          * balancers, see [How Network Load Balancing
-         * Works](https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/overview.htm). This
-         * value is true by default.
+         * Works](https://docs.cloud.oracle.com/Content/Balance/Concepts/balanceoverview.htm#how-network-load-balancing-works).
+         * This value is true by default.
          *
          * <p>Example: {@code true}
          *
@@ -624,8 +623,8 @@ public final class CreateNetworkLoadBalancerDetails
      * list rules](https://docs.cloud.oracle.com/Content/network/Concepts/securitylists.htm) for
      * your virtual cloud network. For more information about public and private network load
      * balancers, see [How Network Load Balancing
-     * Works](https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/overview.htm). This value is
-     * true by default.
+     * Works](https://docs.cloud.oracle.com/Content/Balance/Concepts/balanceoverview.htm#how-network-load-balancing-works).
+     * This value is true by default.
      *
      * <p>Example: {@code true}
      */
@@ -643,8 +642,8 @@ public final class CreateNetworkLoadBalancerDetails
      * list rules](https://docs.cloud.oracle.com/Content/network/Concepts/securitylists.htm) for
      * your virtual cloud network. For more information about public and private network load
      * balancers, see [How Network Load Balancing
-     * Works](https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/overview.htm). This value is
-     * true by default.
+     * Works](https://docs.cloud.oracle.com/Content/Balance/Concepts/balanceoverview.htm#how-network-load-balancing-works).
+     * This value is true by default.
      *
      * <p>Example: {@code true}
      *